+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v.3.0.9
+- fixed coverage report issues
+- fixed and re-factored how tests works
+
 # v.3.0.7
   
  - minor fixes